+++ client/helpers/handlebars.js
@@ -3,7 +3,9 @@
 Handlebars.registerHelper('getSetting', function(setting, defaultArgument){
   return getSetting(setting, defaultArgument);
 });
-
+Handlebars.registerHelper('userProfileRoute', function(displayName){
+  return "/users/"+slugify(displayName);
+});
 Handlebars.registerHelper('canView', function(action) {
   var action=(typeof action !== 'string') ? null : action;
   return canView(Meteor.user(), action);